+ Mensajes en pantalla. + Formato de README.md

+++ lib/rn/commands/books.rb
@@ -18,11 +18,10 @@ module RN
             if name['/'] or name['\\']
               puts "El nuevo nombre tiene caracteres invalidos -> / \\"
             else
-              Dir.mkdir("#{Dir.home}/.my_rns/#{name}/")       
+              Dir.mkdir("#{Dir.home}/.my_rns/#{name}/")
+              puts "El cuaderno se creó correctamente en #{Dir.home}/.my_rns/"
             end
-          end
-          warn "TODO: Implementar creación del cuaderno de notas con nombre '#{name}'.\nPodés comenzar a hacerlo en #{__FILE__}:#{__LINE__}."
-          
+          end          
         end
       end
 
@@ -41,17 +40,23 @@ module RN
         def call(name: nil, **options)
           global = options[:global]
           name = "global" if global
+          if name
             if Dir.exist?("#{Dir.home}/.my_rns/#{name}/")
               if not Dir.empty?("#{Dir.home}/.my_rns/#{name}/")
                 Dir.each_child("#{Dir.home}/.my_rns/#{name}/") {|f| File.delete("#{Dir.home}/.my_rns/#{name}/#{f}")}  
               end
-              Dir.rmdir("#{Dir.home}/.my_rns/#{name}/") if not global
-              puts "El cuaderno se eliminó correctamente"
+              if not global
+                Dir.rmdir("#{Dir.home}/.my_rns/#{name}/") 
+                puts "El cuaderno se eliminó correctamente"
+              else
+                puts "Las notas de global se eliminaron correctamente"
+              end
             else
               puts "El directorio no existe"
             end
-
-          warn "TODO: Implementar borrado del cuaderno de notas con nombre '#{name}' (global=#{global}).\nPodés comenzar a hacerlo en #{__FILE__}:#{__LINE__}."
+          else
+            puts "Se debe proveer un nombre de cuaderno"
+          end
         end
       end
 
@@ -63,8 +68,11 @@ module RN
         ]
 
         def call(*)
-          Dir.each_child("#{Dir.home}/.my_rns/") {|f| puts f}
-          warn "TODO: Implementar listado de los cuadernos de notas.\nPodés comenzar a hacerlo en #{__FILE__}:#{__LINE__}."
+          if Dir.empty?("#{Dir.home}/.my_rns/")
+            puts "No hay cuadernos para mostrar"
+          else
+            Dir.each_child("#{Dir.home}/.my_rns/") {|f| puts f}
+          end
         end
       end
 
@@ -85,12 +93,12 @@ module RN
             if new_name['/'] or new_name['\\'] or new_name[' ']
               puts "El nuevo nombre tiene caracteres invalidos -> '/' '\\' ' '"
             else
-              File.rename("#{Dir.home}/.my_rns/#{old_name}", "#{Dir.home}/.my_rns/#{new_name}")        
+              File.rename("#{Dir.home}/.my_rns/#{old_name}", "#{Dir.home}/.my_rns/#{new_name}")       
+              puts "El cuaderno se renombró correctamente" 
             end    
           else
             puts "El directorio no existe"
           end
-          warn "TODO: Implementar renombrado del cuaderno de notas con nombre '#{old_name}' para que pase a llamarse '#{new_name}'.\nPodés comenzar a hacerlo en #{__FILE__}:#{__LINE__}."
         end
       end
     end
